feat(plugins): route form-builder admin UI through @nextlyhq/ui + SDK (B1/D68)
Move FormLabelWithTooltip into @nextlyhq/ui (re-exported from @nextlyhq/admin
for back-compat), repoint form-builder's component-registry import to
@nextlyhq/plugin-sdk/admin and its FormLabelWithTooltip imports to @nextlyhq/ui.
form-builder/src now has zero @nextlyhq/admin imports (M6 honesty test passes).
Document @nextlyhq/ui as the semver-protected plugin UI kit (D68) in STABILITY.md
and docs/plugins/admin-ui.mdx.
