+++ packages/plugin-form-builder/src/admin/components/builder/ConditionalLogicEditor.tsx
@@ -12,8 +12,8 @@
 
 "use client";
 
-import { FormLabelWithTooltip } from "@nextlyhq/admin";
 import {
+  FormLabelWithTooltip,
   Input,
   Button,
   Checkbox,
