+++ packages/plugin-form-builder/src/admin/components/builder/FieldEditor.tsx
@@ -11,8 +11,8 @@
  */
 
 "use client";
-import { FormLabelWithTooltip } from "@nextlyhq/admin";
 import {
+  FormLabelWithTooltip,
   Input,
   Button,
   Checkbox,
